requirements.txt
streamlit>=1.22.0
pandas>=1.3.5
openpyxl>=3.0.10
requests>=2.28.1
cryptography>=38.0.4
concurrent-futures>=3.0.5
python-dotenv>=0.21.0
tqdm>=4.64.1
pillow>=9.0.0
textwrap3>=0.9.2 